+++ src/wolframite/tools/clerk_helper.clj
@@ -7,8 +7,12 @@
             [clojure.java.io :as io]
             [wolframite.tools.hiccup :as h]))
 
-(defn view [form & {:keys [folded?]}]
+(defn show [form & {:keys [folded?]}]
   (clerk/html (h/view* form folded?)))
+
+(def ^:deprecated view
+  "DEPRECATED - for backwards compatibility with Wolframite <= v1.0.1"
+  show)
 
 (defn stream->bytes [is]
   (let [baos (java.io.ByteArrayOutputStream.)]