+++ src/wolframite/tools/portal.clj
@@ -3,6 +3,10 @@
   (:require [portal.api :as p]
             [wolframite.tools.hiccup :as h]))
 
-(defn view [form & {:keys [folded?]}]
+(defn show [form & {:keys [folded?]}]
   (p/submit (with-meta (h/view* form folded?)
               {:portal.viewer/default :portal.viewer/hiccup})))
+
+(def ^:deprecated view
+  "DEPRECATED - for backwards compatibility with Wolframite <= v1.0.1"
+  show)